Skip directory sync on windows
Windows doesn't support syncing directories. This commit adds a no-op
method for syncing directories on windows.
See
https://github.com/dgraph-io/badger/issues/699#issuecomment-504133587
for more details.
